add senses

+++ CamHi/CommonController/Camera/Setting/DeviceTime/DeviceTimeViewController.m
@@ -229,7 +229,7 @@
 
 - (void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex {
     
-    NSLog(@"buttonIndex:%ld", buttonIndex);
+    NSLog(@"buttonIndex:%d", (int)buttonIndex);
     
     if (buttonIndex == 1) {
         [self.camera request:HI_P2P_SET_TIME_ZONE dson:[self.camera dic:_timeZone]];
